requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-slugify>=8.0.1
aiofiles>=23.2.1
jinja2>=3.1.2
pytest>=7.4.0
httpx>=0.24.1 